refactor: extract GitHubExportProvider+Preflight (#639 slice F)
Sixth slice of #639 — smallest. Moves the pre-flight validation
pair into a dedicated extension file.

Moved (2 methods):
- func validate(configuration:) — already public, unchanged
- makeRepositoryValidationRequest(configuration:) — stays private (single-caller validate moves with)

Zero visibility widenings. validate uses only 'session' (already
widened by #880).

Line-count delta:
- GitHubExportProvider.swift: 272 → 230 (-42 lines)
- GitHubExportProvider+Preflight.swift (new): 47 lines
- xcodeproj: +4 (xcodegen)

Test coverage: GitHubExportProviderTests passes unchanged. External
caller ExportService.swift:41 (validateGitHubConfiguration path)
continues to resolve validate() via extension.

Closes #886. Continues #639 after PRs #875/#879/#881/#883/#885.
